requirements.txt
anthropic>=0.18.0
google-generativeai>=0.3.0
cryptography>=41.0.0
pathlib
flask>=2.3.0
flask-cors>=4.0.0